// src/app/features/shipments/store/shipment-action-types.enum.ts
export enum ShipmentActionTypes {
  LoadShipments = '[Shipments] Load Shipments',
  LoadShipmentsSuccess = '[Shipments] Load Shipments Success',
  LoadShipmentsFailure = '[Shipments] Load Shipments Failure',
  CreateShipment = '[Shipments] Create Shipment',
  CreateShipmentSuccess = '[Shipments] Create Shipment Success',
  CreateShipmentFailure = '[Shipments] Create Shipment Failure',
  CheckoutShipment = '[Shipments] Checkout Shipment',
  CheckoutShipmentSuccess = '[Shipments] Checkout Shipment Success',
  CheckoutShipmentFailure = '[Shipments] Checkout Shipment Failure',
  DeliverShipment = '[Shipments] Deliver Shipment',
  DeliverShipmentSuccess = '[Shipments] Deliver Shipment Success',
  DeliverShipmentFailure = '[Shipments] Deliver Shipment Failure',
  DeleteShipment = '[Shipments] Delete Shipment',
  DeleteShipmentSuccess = '[Shipments] Delete Shipment Success',
  DeleteShipmentFailure = '[Shipments] Delete Shipment Failure'
}